Allowed the bot to be able to be tested while the main bot is running
Allowed the bot to be able to be tested while the main bot is running by having the test channel ids be defined when `TEST` is defined with `1` in a environment file.

Also added application commands error catcher and also a global bot command checker so commands can only be run in PMAM servers.