Changed logger to rotate log file 3 minutes after midnight
Changed logger to rotate log file 3 minutes after midnight because at midnight the bot will reboot. This should give some time for the bot to restart and start up again. Why 3 instead of something earlier is to also account for the restart service that occurs to reboot the server every Sunday. The server should in its entirety should be up again before 00:03 to start up the bot again and rotate the log file. If it for some reason is past this time, setupLogging will check if its past the time and manually rotate.