Redone how the on_message_delete event works.
The event now should not fail to get deleted attachments, but the main change is that the whole event has been reworked. The event now gets all attachments and logs them when they are deleted. If the file is over 8MB it gets compressed to 7zip format before being uploaded, but if even that is over 8MB still, then it's stored locally on the server for 30 days which then gets automatically deleted.

dotenv module has been re-added so debugging can be possible without having to execute `./runbot.sh`. This also makes it easier to debug on Windows.