Redone the Mods/Admin `reply` command, now named `modmail`.
Redone the `reply` command now renamed to `modmail` so that it properly supports attachments. There are two versions, the command and slash command form. Each with its own pros and cons, but key difference is that one is a interaction with a modal while the other is simply a command with a prefix.

Still has some trouble sending other files other than images but still works pretty well.